Refactor Ecto module: Remove duplicate code and add default function
